+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dto/DeveloperACBMapDTO.java
@@ -2,8 +2,8 @@ package gov.healthit.chpl.dto;
 
 import java.io.Serializable;
 
-import gov.healthit.chpl.entity.DeveloperACBMapEntity;
-import gov.healthit.chpl.entity.DeveloperACBTransparencyMapEntity;
+import gov.healthit.chpl.entity.developer.DeveloperACBMapEntity;
+import gov.healthit.chpl.entity.developer.DeveloperACBTransparencyMapEntity;
 
 public class DeveloperACBMapDTO implements Serializable {
 	private static final long serialVersionUID = -1860729017532925654L;