+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/entity/ProductActiveOwnerEntity.java
@@ -14,6 +14,8 @@ import javax.persistence.JoinColumn;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
 
+import gov.healthit.chpl.entity.developer.DeveloperEntity;
+
 
 @Entity
 @Table(name = "product_active_owner_history_map")