+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/entity/developer/DeveloperStatusEntity.java
@@ -1,4 +1,4 @@
-package gov.healthit.chpl.entity;
+package gov.healthit.chpl.entity.developer;
 
 import java.io.Serializable;
 import java.util.Date;
@@ -26,7 +26,7 @@ public class DeveloperStatusEntity implements Cloneable, Serializable {
 	private Long id;
 	
 	@Column(name = "name")
-	@Type(type = "gov.healthit.chpl.entity.PostgresDeveloperStatusType" , parameters ={@org.hibernate.annotations.Parameter(name = "enumClassName",value = "gov.healthit.chpl.entity.DeveloperStatusType")} )
+	@Type(type = "gov.healthit.chpl.entity.developer.PostgresDeveloperStatusType" , parameters ={@org.hibernate.annotations.Parameter(name = "enumClassName",value = "gov.healthit.chpl.entity.developer.DeveloperStatusType")} )
 	private DeveloperStatusType name;
 
 	@Basic( optional = false )