+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/entity/developer/DeveloperStatusType.java
@@ -1,4 +1,4 @@
-package gov.healthit.chpl.entity;
+package gov.healthit.chpl.entity.developer;
 
 import org.springframework.util.StringUtils;