+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/impl/ProductVersionManagerImpl.java
@@ -31,8 +31,8 @@ import gov.healthit.chpl.dto.DeveloperDTO;
 import gov.healthit.chpl.dto.DeveloperStatusEventDTO;
 import gov.healthit.chpl.dto.ProductDTO;
 import gov.healthit.chpl.dto.ProductVersionDTO;
-import gov.healthit.chpl.entity.DeveloperStatusType;
 import gov.healthit.chpl.entity.ProductVersionEntity;
+import gov.healthit.chpl.entity.developer.DeveloperStatusType;
 import gov.healthit.chpl.manager.ActivityManager;
 import gov.healthit.chpl.manager.ProductVersionManager;